requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.2
kiteconnect==4.2.0
pandas==2.1.3
numpy==1.26.2
pytest==7.4.3
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
